refactor(tracking): store helper functions in a helper folder
