+++ packages/orbit-tracking/src/helpers/category.ts
@@ -1,0 +1,11 @@
+import { CATEGORIES } from "../consts";
+
+export const getCategory = (name: string) => {
+  for (const [cat, components] of Object.entries(CATEGORIES)) {
+    if (components.includes(name)) {
+      return cat.toLowerCase();
+    }
+  }
+
+  return "unknown";
+};
