+++ packages/orbit-tracking/src/helpers/message.ts
@@ -1,0 +1,4 @@
+import chalk from "chalk";
+
+export const errorMessage = (str: string) => console.error(chalk.red.bold.underline(str));
+export const infoMessage = (str: string) => console.info(chalk.magenta.bold(str));
