+++ packages/orbit-tracking/src/helpers/path.ts
@@ -1,0 +1,10 @@
+import { fileURLToPath } from "url";
+import path from "path";
+
+const __filename = fileURLToPath(import.meta.url);
+export const __dirname = path.dirname(__filename);
+
+export const getOutputPath = (p: string, name: string) => {
+  if (p) return path.resolve(p, `tracking-${name}.json`);
+  return path.resolve(process.cwd(), `tracking-${name}.json`);
+};
