+++ packages/orbit-tracking/src/helpers/template.ts
@@ -1,0 +1,9 @@
+export const headingTemplate = (str: string) => `
+  ++++++++++++++++ ++++++++++++++++
+  ++++++++++++++++ ++++++++++++++++
+
+  ${str}
+
+  ++++++++++++++++ ++++++++++++++++
+  ++++++++++++++++ ++++++++++++++++
+`;
