+++ packages/orbit-tracking/src/helpers/timestamp.ts
@@ -1,0 +1,7 @@
+export const timestamp = () => {
+  const date = new Date().getDate();
+  const month = new Date().getMonth() + 1;
+  const year = new Date().getFullYear();
+
+  return `${date}-${month}-${year}`;
+};
