+++ packages/orbit-tracking/src/index.ts
@@ -1,8 +1,8 @@
-import queries from "./api";
+import { queries, mutations } from "./api";
 import cli from "./cli";
 import fetcher from "./fetcher";
 import orbitProcessor from "./orbitProcessor";
 import * as helpers from "./helpers";
 import * as consts from "./consts";
 
-export { cli, fetcher, orbitProcessor, helpers, consts, queries };
+export { cli, fetcher, orbitProcessor, helpers, consts, queries, mutations };
